requirements.txt
Flask
numpy
pandas
scikit-learn
joblib
nltk
wtforms
newspaper3k
requests
gunicorn